Update to Rails 4.2.6 and ActiveJob
Active Job broke the tutorial, so I updated the code so it works.

…sort of.

Deserialization of Global Identifiers was added too! It’s awesome, but
it means that the post id needs to be stored in not-an-instance
variable because deleting a document uses the post id, and deleting it
immediately kinda means that the post id won’t exist when the job runs.

Creating documents works just fine though.

Miiiiight want to fix the whole deletion not working thing.